react: proxy exports
