+++ packages/node/package.json
@@ -1,7 +1,13 @@
 {
   "name": "@renegade-fi/node",
-  "version": "0.0.34",
   "description": "Node.js library for Renegade",
+  "scripts": {
+    "build": "pnpm run clean && pnpm run build:esm+types",
+    "build:esm+types": "tsc --project tsconfig.build.json --outDir ./dist --declaration --declarationMap --declarationDir ./dist/types",
+    "clean": "rm -rf dist tsconfig.tsbuildinfo",
+    "typecheck": "tsc --noEmit"
+  },
+  "version": "0.0.34",
   "files": [
     "dist/**",
     "!dist/**/*.tsbuildinfo",
@@ -22,17 +28,9 @@
     },
     "./package.json": "./package.json"
   },
-  "scripts": {
-    "build": "pnpm run clean && pnpm run build:esm+types",
-    "build:esm+types": "tsc --project tsconfig.build.json --outDir ./dist --declaration --declarationMap --declarationDir ./dist/types",
-    "clean": "rm -rf dist tsconfig.tsbuildinfo",
-    "typecheck": "tsc --noEmit"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
     "@renegade-fi/core": "workspace:*",
     "viem": "2.9.28"
-  }
+  },
+  "keywords": []
 }
