+++ packages/react/src/exports/constants.ts
@@ -1,0 +1,5 @@
+////////////////////////////////////////////////////////////////////////////////
+// @renegade-fi/core/constants
+////////////////////////////////////////////////////////////////////////////////
+
+export * from '@renegade-fi/core/constants'
